chore: refactor business structure endpoint
